Use Elasticsearch for repository
Instead of keeping an immutable in-memory list of posts,
use Elasticsearch. This is set up to run on a local node,
persisting data to /tmp which works fine for development;
this should be changed to point at an actual cluster
(docker container, web service, whatever) for production.

To initialise this with index 'boogle' and type 'post', run
client.execute {
  createIndex("boogle").mappings(
    mapping("post").fields(
      textField("title"),
      textField("body")
    )
  )
}

Test data was added to mimic the initial data contained in the list.
Request mappings work, start the app with 'sbt run' in the shell.

Lots of refactoring to be done.