Change Post into Book
Change the Post JSON object used by the API into a book, with:
* Title
* Author
* List of pages (assumed to be complete and in order,
  with page n at index n-1

Store book data (title and author) in one index, and pages
(book ID, page number and content) in another.
Implement retrieval of book (and pages) by ID or list.

This has been tested with each API method so it works,
but there's a lot of refactoring to do - main points:
* See if we should be making blocking 'await' calls inside
  futures used by Repository, it feels wrong. Find best way
  to do this in Scala - Java way might be to chain Futures.
* Check if we can use the elastic4s API to retrieve books and
  pages by ID, rather than doing a search without a query
  then filtering in the Repository.
* Change Elasticsearch index setup - is there the equivalent
  of a 'foreign key' from page to book, or some kind of parent
  -child relationship? Although each field is indexed so this
  shouldn't matter much for performance,
  it would simplify updates.